requirements.txt
streamlit
pandas
fpdf2
python-docx
Pillow
streamlit-drawable-canvas
numpy
streamlit-option-menu